+++ scripts/gwf
@@ -5,11 +5,11 @@ import os.path
 import sys
 import argparse
 import pkg_resources
-import mk_gwf_workflow
-from mk_gwf_workflow.backends import AVAILABLE_BACKENDS
-from mk_gwf_workflow.configurations import read_configurations
-from mk_gwf_workflow.colours import *
-from mk_gwf_workflow.performance import *
+import gwf_workflow
+from gwf_workflow.backends import AVAILABLE_BACKENDS
+from gwf_workflow.configurations import read_configurations
+from gwf_workflow.colours import *
+from gwf_workflow.performance import *
 
 # This gets the installed version of gwf that is actually being used by the script.
 package_version = pkg_resources.require("gwf")[0].version
@@ -54,11 +54,11 @@ if not os.path.exists(args.workflow_file):
 
 
 # Set the backend based on defaults or arguments...
-mk_gwf_workflow.BACKEND = AVAILABLE_BACKENDS[args.backend]()
+gwf_workflow.BACKEND = AVAILABLE_BACKENDS[args.backend]()
 
 
 execfile(args.workflow_file)
-from mk_gwf_workflow.workflow import build_workflow, schedule, dependencies
+from gwf_workflow.workflow import build_workflow, schedule, dependencies
 workflow = build_workflow()
 
 if args.performance:
