+++ src/gwf/bwa.py
